LayoutTests/fast/shadow-dom/cannot-hit-test-into-shadow-file-input-expected.txt

PASS range.commonAncestorContainer is fileInputElement
PASS successfullyParsed is true

TEST COMPLETE
